pyproject.toml
[tool.poetry]
name = "rag-app-3"
version = "0.1.0"
description = ""
authors = ["Shubehndu Shubham <s2.shubh2@gmail.com>"]
readme = "README.md"
packages = [{include = "rag_app_3"}]

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
langchain-anthropic = "^0.2.1"
langchain-ollama = "^0.2.0"
ipykernel = "^6.29.5"
python-dotenv = "^1.0.1"
streamlit = "^1.38.0"
langchain-community = "^0.3.0"
transformers = "^4.44.2"
tqdm = "^4.66.5"
torch = "^2.4.1"
accelerate = "^0.34.2"
langchain-openai = "^0.2.0"
streamlit-lottie = "^0.0.5"
boto3 = "^1.35.24"
awscli = "^1.34.24"
unstructured = {extras = ["all-docs"], version = "^0.15.13"}
onnx = "1.16.1"
langchain-experimental = "^0.3.0"
langchain-groq = "^0.2.0"
langchain-mistralai = "^0.2.0"
langchain-huggingface = "^0.1.0"
python-magic-bin = "^0.4.14"
pypdf2 = "^3.0.1"
docx2txt = "^0.8"
faiss-cpu = "^1.8.0.post1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
